Switch to libedit-dev from libreadline6-dev
libreadline is under the GPL with no linking exceptions. This makes
linking with software under the PHP license (among others, like e.g.
OpenSSL as well) undistributable. Therefore, switch to the BSD-licensed
libedit-dev to avoid this pitfall.

To do so, add patch debian/patches/prefer-libedit to workaround a build
system bug that would result in a FTBFS. The patch description has all
the details.

Finally, add a debian/TODO item to review other incompatibilities such
as this one, as the list of direct and indirect dependencies is large
and the possibility of GPL-only software creeping in is considerable.