Add Provides:hhvm-api-HHVM_API_VERSION
HHVM has a sort-of API/ABI compatibility number in the HHVM_API_VERSION
define. To make it easier for extensions packagers to provide a correct
dependency we add a Provides: hhvm-api-$version to the hhvm
package. Also, changing the API_VERSION can allow packagers of hhvm
itself to indicate extensions packagers when to forcibly rebuild their
packages.