[WMF] Add PCRE patches to fix our throughput problems
The PCRE cache in HHVM 3.3.0 is pretty naive, and will grow indefinitely
until the maximum cache size is reached. Also, compiling of pcre is done
sequentially. This wouldn't be an issue if we didn't generate an
indefinite amount of regexes, which is the case for mediawiki.

This patch (which is a bundle of 4 upstream patches) fixes the issue.

Change-Id: Ia42be8335958700e3b847aca91aa2e2a38768400